Assert public domain license on arrangements of traditional tunes.
Most, if not all of the non-original tunes in this repo will be
arrangements of traditional tunes in the public domain.

There are no non-original tunes with other copyright holders now;
let's cross that bridge when we come to it.

IANAL, so if anybody has better suggestions for the LICENSE file,
patches welcome.